resolve gkm conflicts
